Print the average each iteration when looping forever
Otherwise it doesn't display an average at all.  It is better to
display it every loop because your eye can easily spot the average
based on the one that doesn't change as much.
Press return to reset the average, I find on one system killing and
restarting can vary a lot, better to keep it running and reset the
average when something else changed on the system.